move neural network engine to separate file
